# src/generated/service.openapi.json
{"openapi":"3.0.2","info":{"title":"customer-address-service","description":"stores customer addresses","version":"1.0.3"},"servers":[{"url":"https://api.eu-central-1.dev.services.vwfs.io/customer-address-service"},{"url":"https://api.eu-central-1.int.services.vwfs.io/customer-address-service"},{"url":"https://api.eu-central-1.cons.services.vwfs.io/customer-address-service"},{"url":"https://api.eu-central-1.services.vwfs.io/customer-address-service"},{"url":"https://api.us-east-1.dev.services.vwfs.io/customer-address-service"},{"url":"https://api.us-east-1.int.services.vwfs.io/customer-address-service"},{"url":"https://api.us-east-1.cons.services.vwfs.io/customer-address-service"},{"url":"https://api.us-east-1.services.vwfs.io/customer-address-service"}],"paths":{"/customers/{customerId}/addresses":{"post":{"tags":["CustomerAddressManagement"],"operationId":"createCustomerAddress","summary":"creates new customer","description":"creates a new customer with a specified address","parameters":[{"name":"customerId","in":"path","description":"Id of the specified customer","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressPostRequest"}}},"required":true},"responses":{"201":{"description":"Returns the stored customer address and its identifier","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressPostResponse"}}}}},"x-codegen-request-body-name":"CustomerAddressPostRequest","security":[{"ApiKeyAuth":[]}]},"get":{"tags":["CustomerAddressManagement"],"operationId":"getCustomerAddresses","summary":"Get Customer Addresses","description":"Gets the Adresseses for the specified customer","parameters":[{"name":"customerId","in":"path","description":"Id of the specified customer","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"the customer address of the specified customer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressesGetResponse"}}}},"404":{"description":"Resource not found","content":{"application/json":{"schema":{"additionalProperties":true}}}}},"security":[{"ApiKeyAuth":[]}]}},"/customers/{customerId}/addresses/{addressId}":{"delete":{"tags":["CustomerAddressManagement"],"operationId":"deleteCustomerAddresses","summary":"deletes Customer Addresses","description":"deletes an address for the specified customer","parameters":[{"name":"customerId","in":"path","description":"Id of the specified customer","required":true,"schema":{"type":"string"}},{"name":"addressId","in":"path","description":"Id of the specified address","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"resource deleted successfully","content":{"application/json":{"schema":{"additionalProperties":true}}}},"404":{"description":"resource not found","content":{"application/json":{"schema":{"additionalProperties":true}}}}},"security":[{"ApiKeyAuth":[]}]}},"/customers/{customerId}/addresses/{addressId}/verification":{"post":{"tags":["CustomerAddressVerification"],"operationId":"createVerificationDocument","summary":"uploads verification document","description":"uploads a verification document and saves it in an S3 bucket.","parameters":[{"name":"customerId","in":"path","description":"Id of the specified customer","required":true,"x-obfuscate":false,"schema":{"type":"string"}},{"name":"addressId","in":"path","description":"Id of the specified customer","required":true,"x-obfuscate":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressVerificationRequest"}}},"required":true},"responses":{"201":{"description":"The customer id with a conformation that the document has been saved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressPostResponse"}}}},"404":{"description":"Could not find the customer or the address for that specific customer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressNotFound"}}}}},"security":[{"ApiKeyAuth":[]}]},"get":{"tags":["CustomerAddressVerification"],"operationId":"getVerificationDocument","summary":"gets verification Document","description":"checks if customer ","parameters":[{"name":"customerId","in":"path","description":"Id of the specified customer","required":true,"x-obfuscate":false,"schema":{"type":"string"}},{"name":"addressId","in":"path","description":"Id of the specified customer","required":true,"x-obfuscate":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"the verification document as base64","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerAddressVerificationGetResponse"}}}}},"security":[{"ApiKeyAuth":[]}]}}},"components":{"securitySchemes":{"ApiKeyAuth":{"type":"apiKey","name":"x-api-key","in":"header"}},"schemas":{"CustomerAddressesGetResponse":{"type":"array","items":{"type":"object","properties":{"addressId":{"type":"string","description":"the unique Id of the address","example":"879b57a7-843a-41d8-991b-7970c8431594"},"address":{"type":"object","description":"the address of a customer","allOf":[{"$ref":"#/components/schemas/Address"}]}}}},"CustomerAddressPostRequest":{"type":"object","properties":{"address":{"type":"object","description":"the address of a customer","allOf":[{"$ref":"#/components/schemas/Address"}]}},"required":["address"]},"CustomerAddressPostResponse":{"type":"object","properties":{"addressId":{"type":"string","description":"a unique identifier of the newly persisted address","example":"3fbb832a-28b9-4f4d-8e2a-3bf65c50c2e6"}},"required":["addressId"]},"CustomerAddressVerificationRequest":{"type":"object","allOf":[{"$ref":"#/components/schemas/CustomerAddressVerificationDocument"}]},"CustomerAddressVerificationGetResponse":{"type":"object","allOf":[{"$ref":"#/components/schemas/CustomerAddressVerificationDocument"}]},"CustomerAddressVerificationDocument":{"type":"object","properties":{"base64":{"type":"string","description":"verification document as base64 encoded string","example":"YWxidXM="}},"required":["base64"]},"CustomerAddressNotFound":{"type":"object","properties":{"customerId":{"type":"string","description":"An ID of the customer"},"addressId":{"type":"string","description":"An ID of the address"},"message":{"type":"string","description":"An error message"}}},"Address":{"type":"object","properties":{"street":{"type":"string","description":"The street of the address.","example":"Goethestraße"},"houseNumber":{"type":"string","description":"The house number of the address including extra characters like in \"29a\".","example":"29a"},"optionalLine":{"type":"string","description":"This contains any additional information regarding location of the address at the specified housenumber","example":"Block C, staircase 5, LEFT"},"zipCode":{"type":"string","description":"The zip code of the address.","example":"10162"},"city":{"type":"string","description":"The city of the address.","example":"Berlin"},"countryCode":{"title":"enum","description":"The country code of the country where the address is located.","type":"string","example":"DE","allOf":[{"$ref":"#/components/schemas/CountryCode"}]},"residentSinceDate":{"type":"string","description":"The date since when a certain person is living at the address. This should only be filled if the resident is a person.","example":"2000-11-20","pattern":"^(-?(?:[1-9][0-9]*)?[0-9]{4})-(1[0-2]|0[1-9])-(3[01]|0[1-9]|[12][0-9])(T(2[0-3]|[01][0-9]):([0-5][0-9]):([0-5][0-9])(\\.[0-9]+)?(Z)?)?$"},"isMainAddress":{"type":"boolean","description":"A flag that can be used to mark an address as the main one when multiple addresses are available e.g. in customer data.","example":true}},"required":["street","houseNumber","zipCode","city","countryCode"]},"CountryCode":{"description":"Represents an ISO 3166-ALPHA-2 conform country code accessible through the full english name of the country. This enum should be used to describe for instance vehicle origins or nationalities of persons. Don't use it to describe a VWFS market. Use enum Country in @bus/models.","enum":["AF","AX","AL","DZ","AS","AD","AO","AI","AQ","AG","AR","AM","AW","AU","AT","AZ","BS","BH","BD","BB","BY","BE","BZ","BJ","BM","BT","BO","BA","BW","BV","BR","IO","BN","BG","BF","BI","KH","CM","CA","CV","KY","CF","TD","CL","CN","CX","CC","CO","KM","CG","CD","CK","CR","HR","CU","CY","CZ","DK","DJ","DM","DO","EC","EG","SV","GQ","ER","EE","ET","FK","FO","FJ","FI","FR","GF","PF","TF","GA","GM","GE","DE","GH","GI","GR","GL","GD","GP","GU","GT","GG","GN","GW","GY","HT","HM","VA","HN","HK","HU","IS","IN","ID","IR","IQ","IE","IM","IL","IT","CI","JM","JP","JE","JO","KZ","KE","KI","KR","KW","KG","LA","LV","LB","LS","LR","LY","LI","LT","LU","MO","MK","MG","MW","MY","MV","ML","MT","MH","MQ","MR","MU","YT","MX","FM","MD","MC","MN","ME","MS","MA","MZ","MM","NA","NR","NP","NL","AN","NC","NZ","NI","NE","NG","NU","NF","MP","NO","OM","PK","PW","PS","PA","PG","PY","PE","PH","PN","PL","PT","PR","QA","RE","RO","RU","RW","BL","SH","KN","LC","MF","PM","VC","WS","SM","ST","SA","SN","RS","SC","SL","SG","SK","SI","SB","SO","ZA","GS","ES","LK","SD","SR","SJ","SZ","SE","CH","SY","TW","TJ","TZ","TH","TL","TG","TK","TO","TT","TN","TR","TM","TC","TV","UG","UA","AE","GB","US","UM","UY","UZ","VU","VE","VN","VG","VI","WF","EH","YE","ZM","ZW"],"x-enum":{"Afghanistan":{"value":"AF"},"AlandIslands":{"value":"AX"},"Albania":{"value":"AL"},"Algeria":{"value":"DZ"},"AmericanSamoa":{"value":"AS"},"Andorra":{"value":"AD"},"Angola":{"value":"AO"},"Anguilla":{"value":"AI"},"Antarctica":{"value":"AQ"},"AntiguaAndBarbuda":{"value":"AG"},"Argentina":{"value":"AR"},"Armenia":{"value":"AM"},"Aruba":{"value":"AW"},"Australia":{"value":"AU"},"Austria":{"value":"AT"},"Azerbaijan":{"value":"AZ"},"Bahamas":{"value":"BS"},"Bahrain":{"value":"BH"},"Bangladesh":{"value":"BD"},"Barbados":{"value":"BB"},"Belarus":{"value":"BY"},"Belgium":{"value":"BE"},"Belize":{"value":"BZ"},"Benin":{"value":"BJ"},"Bermuda":{"value":"BM"},"Bhutan":{"value":"BT"},"Bolivia":{"value":"BO"},"BosniaAndHerzegovina":{"value":"BA"},"Botswana":{"value":"BW"},"BouvetIsland":{"value":"BV"},"Brazil":{"value":"BR"},"BritishIndianOceanTerritory":{"value":"IO"},"BruneiDarussalam":{"value":"BN"},"Bulgaria":{"value":"BG"},"BurkinaFaso":{"value":"BF"},"Burundi":{"value":"BI"},"Cambodia":{"value":"KH"},"Cameroon":{"value":"CM"},"Canada":{"value":"CA"},"CapeVerde":{"value":"CV"},"CaymanIslands":{"value":"KY"},"CentralAfricanRepublic":{"value":"CF"},"Chad":{"value":"TD"},"Chile":{"value":"CL"},"China":{"value":"CN"},"ChristmasIsland":{"value":"CX"},"CocosKeelingIslands":{"value":"CC"},"Colombia":{"value":"CO"},"Comoros":{"value":"KM"},"Congo":{"value":"CG"},"CongoDemocraticRepublic":{"value":"CD"},"CookIslands":{"value":"CK"},"CostaRica":{"value":"CR"},"Croatia":{"value":"HR"},"Cuba":{"value":"CU"},"Cyprus":{"value":"CY"},"CzechRepublic":{"value":"CZ"},"Denmark":{"value":"DK"},"Djibouti":{"value":"DJ"},"Dominica":{"value":"DM"},"DominicanRepublic":{"value":"DO"},"Ecuador":{"value":"EC"},"Egypt":{"value":"EG"},"ElSalvador":{"value":"SV"},"EquatorialGuinea":{"value":"GQ"},"Eritrea":{"value":"ER"},"Estonia":{"value":"EE"},"Ethiopia":{"value":"ET"},"FalklandIslands":{"value":"FK"},"FaroeIslands":{"value":"FO"},"Fiji":{"value":"FJ"},"Finland":{"value":"FI"},"France":{"value":"FR"},"FrenchGuiana":{"value":"GF"},"FrenchPolynesia":{"value":"PF"},"FrenchSouthernTerritories":{"value":"TF"},"Gabon":{"value":"GA"},"Gambia":{"value":"GM"},"Georgia":{"value":"GE"},"Germany":{"value":"DE"},"Ghana":{"value":"GH"},"Gibraltar":{"value":"GI"},"Greece":{"value":"GR"},"Greenland":{"value":"GL"},"Grenada":{"value":"GD"},"Guadeloupe":{"value":"GP"},"Guam":{"value":"GU"},"Guatemala":{"value":"GT"},"Guernsey":{"value":"GG"},"Guinea":{"value":"GN"},"GuineaBissau":{"value":"GW"},"Guyana":{"value":"GY"},"Haiti":{"value":"HT"},"HeardIslandMcdonaldIslands":{"value":"HM"},"HolySeeVaticanCityState":{"value":"VA"},"Honduras":{"value":"HN"},"HongKong":{"value":"HK"},"Hungary":{"value":"HU"},"Iceland":{"value":"IS"},"India":{"value":"IN"},"Indonesia":{"value":"ID"},"Iran":{"value":"IR"},"Iraq":{"value":"IQ"},"Ireland":{"value":"IE"},"IsleOfMan":{"value":"IM"},"Israel":{"value":"IL"},"Italy":{"value":"IT"},"IvoryCoast":{"value":"CI"},"Jamaica":{"value":"JM"},"Japan":{"value":"JP"},"Jersey":{"value":"JE"},"Jordan":{"value":"JO"},"Kazakhstan":{"value":"KZ"},"Kenya":{"value":"KE"},"Kiribati":{"value":"KI"},"Korea":{"value":"KR"},"Kuwait":{"value":"KW"},"Kyrgyzstan":{"value":"KG"},"LaoPeoplesDemocraticRepublic":{"value":"LA"},"Latvia":{"value":"LV"},"Lebanon":{"value":"LB"},"Lesotho":{"value":"LS"},"Liberia":{"value":"LR"},"LibyanArabJamahiriya":{"value":"LY"},"Liechtenstein":{"value":"LI"},"Lithuania":{"value":"LT"},"Luxembourg":{"value":"LU"},"Macao":{"value":"MO"},"Macedonia":{"value":"MK"},"Madagascar":{"value":"MG"},"Malawi":{"value":"MW"},"Malaysia":{"value":"MY"},"Maldives":{"value":"MV"},"Mali":{"value":"ML"},"Malta":{"value":"MT"},"MarshallIslands":{"value":"MH"},"Martinique":{"value":"MQ"},"Mauritania":{"value":"MR"},"Mauritius":{"value":"MU"},"Mayotte":{"value":"YT"},"Mexico":{"value":"MX"},"Micronesia":{"value":"FM"},"Moldova":{"value":"MD"},"Monaco":{"value":"MC"},"Mongolia":{"value":"MN"},"Montenegro":{"value":"ME"},"Montserrat":{"value":"MS"},"Morocco":{"value":"MA"},"Mozambique":{"value":"MZ"},"Myanmar":{"value":"MM"},"Namibia":{"value":"NA"},"Nauru":{"value":"NR"},"Nepal":{"value":"NP"},"Netherlands":{"value":"NL"},"NetherlandsAntilles":{"value":"AN"},"NewCaledonia":{"value":"NC"},"NewZealand":{"value":"NZ"},"Nicaragua":{"value":"NI"},"Niger":{"value":"NE"},"Nigeria":{"value":"NG"},"Niue":{"value":"NU"},"NorfolkIsland":{"value":"NF"},"NorthernMarianaIslands":{"value":"MP"},"Norway":{"value":"NO"},"Oman":{"value":"OM"},"Pakistan":{"value":"PK"},"Palau":{"value":"PW"},"PalestinianTerritory":{"value":"PS"},"Panama":{"value":"PA"},"PapuaNewGuinea":{"value":"PG"},"Paraguay":{"value":"PY"},"Peru":{"value":"PE"},"Philippines":{"value":"PH"},"Pitcairn":{"value":"PN"},"Poland":{"value":"PL"},"Portugal":{"value":"PT"},"PuertoRico":{"value":"PR"},"Qatar":{"value":"QA"},"Reunion":{"value":"RE"},"Romania":{"value":"RO"},"RussianFederation":{"value":"RU"},"Rwanda":{"value":"RW"},"SaintBarthelemy":{"value":"BL"},"SaintHelena":{"value":"SH"},"SaintKittsAndNevis":{"value":"KN"},"SaintLucia":{"value":"LC"},"SaintMartin":{"value":"MF"},"SaintPierreAndMiquelon":{"value":"PM"},"SaintVincentAndGrenadines":{"value":"VC"},"Samoa":{"value":"WS"},"SanMarino":{"value":"SM"},"SaoTomeAndPrincipe":{"value":"ST"},"SaudiArabia":{"value":"SA"},"Senegal":{"value":"SN"},"Serbia":{"value":"RS"},"Seychelles":{"value":"SC"},"SierraLeone":{"value":"SL"},"Singapore":{"value":"SG"},"Slovakia":{"value":"SK"},"Slovenia":{"value":"SI"},"SolomonIslands":{"value":"SB"},"Somalia":{"value":"SO"},"SouthAfrica":{"value":"ZA"},"SouthGeorgiaAndSandwichIsl":{"value":"GS"},"Spain":{"value":"ES"},"SriLanka":{"value":"LK"},"Sudan":{"value":"SD"},"Suriname":{"value":"SR"},"SvalbardAndJanMayen":{"value":"SJ"},"Swaziland":{"value":"SZ"},"Sweden":{"value":"SE"},"Switzerland":{"value":"CH"},"SyrianArabRepublic":{"value":"SY"},"Taiwan":{"value":"TW"},"Tajikistan":{"value":"TJ"},"Tanzania":{"value":"TZ"},"Thailand":{"value":"TH"},"TimorLeste":{"value":"TL"},"Togo":{"value":"TG"},"Tokelau":{"value":"TK"},"Tonga":{"value":"TO"},"TrinidadAndTobago":{"value":"TT"},"Tunisia":{"value":"TN"},"Turkey":{"value":"TR"},"Turkmenistan":{"value":"TM"},"TurksAndCaicosIslands":{"value":"TC"},"Tuvalu":{"value":"TV"},"Uganda":{"value":"UG"},"Ukraine":{"value":"UA"},"UnitedArabEmirates":{"value":"AE"},"UnitedKingdom":{"value":"GB"},"UnitedStates":{"value":"US"},"UnitedStatesOutlyingIslands":{"value":"UM"},"Uruguay":{"value":"UY"},"Uzbekistan":{"value":"UZ"},"Vanuatu":{"value":"VU"},"Venezuela":{"value":"VE"},"Vietnam":{"value":"VN"},"VirginIslandsBritish":{"value":"VG"},"VirginIslandsUS":{"value":"VI"},"WallisAndFutuna":{"value":"WF"},"WesternSahara":{"value":"EH"},"Yemen":{"value":"YE"},"Zambia":{"value":"ZM"},"Zimbabwe":{"value":"ZW"}}}}}}